+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 2.4.8
+- Fix missing Chicken Pool and Missing Pool warnings in Pumpkin Patch Village
 
 # 2.4.7
 - Add Pale Pumpkin Seeds to Chicken and Parrot Food